State the scripts that use a table.
I needed to move the filename value from VisualNode to EvolutionNode (why was it in Visual? Evolution nodes are to know that information).
I made make a list of the filenames that are using a relation and then pass it to the pane in order to display them in File Structure pane.
I changed the Visualization not to display Pick, Move since these features were not working (evakont).