Display all affected modules @ impact anal tab.
I give @ drawing functions the nodes that are for the module that is to be displayed.
I sort modules to firstly draw the queries/views and then the relations (TODO: check what happens @ chains of views --probably not good result).
Fixed a bug (I don't know why this happens) that sets the graph size to (0,0)! I deribelately set it to (200,200).
Fixed a bug that had to do with the filename @ tooltip and another that had to do with epilegmenos kombos @HecataeusPolicyManagerGUI.