Restructrued Hecataeus GUI.
The summary graph is in left panel, below the file structure.
Next to them is the zoom area, where we have the zoomed modules and the impact analysis tabs, when an event is placed.
Finally, in the rightmost are we have the two panels that have to do with the policies and the events, and below them an text area we use for notifications.