Finished with file structure part.
I display folders with colors and their children are brighter that the folders.
TODO: fix the recolorize() function a little bit.
TODO: children of a folder should have different colors too, the first brighter() the second brighter().brighter() etc. (of course this has to do with the recolorize() function, so I just mention it to remember.