Restructure of drupal project.
I created it as if I had found the queries in the script files, by searching the source code of the project.  I recolorize the folders of the FileStructureGUI to take a color as their parent, but slightly modified.  I also added the main graph to the Zoom area, now the user can interact with both panels in order to find/select/modify nodes (this is not good code, since Evolution and Visual graphs should be separated, but it works till now).  Fixed a bug that occured when a user had oppened a project and then opened another (I didn't clear the FileStructureGUI panel, so there were problems).